+++ CHANGELOG.rst
@@ -21,7 +21,8 @@ Changed
 Added
 ~~~~~
 - added support for arbitrary hashable sequence in the pure Python fallback implementation of all functions in ``rapidfuzz.distance``
-
+- added support for ``None`` and ``float("nan")`` in ``process.cdist`` as long as the underlying scorer supports it.
+  This is the case for all scorers returning normalized results.
 
 [2.15.1] - 2023-04-11
 ^^^^^^^^^^^^^^^^^^^^^